separate data packaging into respective papers
